Test macros clean-up and improvements: - Added Monopole example test program and test macro and including them in test suites - Added coloured output  for test_suite_exe.sh - Do not exclude ExGarfield test with Root 6 - Added script options to other tests as in test_suites.sh - Recreate test output directory only if test is required for   all examples and both MCs - Fixed typo in test programs description and PrintUsage() function - Further simplifications of scripts

+++ examples/run_suite_exe.sh
@@ -9,36 +9,63 @@
 #-------------------------------------------------
 
 #
-# Run tests for all VMC examples from built executables and regenerate output
-# files. Only g4Config.C and g3tgeoConfig.C configurations are tested.
-# The executables are supposed to be found in
-# examples_build_g3/exampleName and examples_build_g4/exampleName
-# directories.
+# Run tests for all VMC examples from built executables.
+# Only g4tgeoConfig.C and g3tgeoConfig.C configurations are tested.
+# The executables are supposed to be on the executable path.
+#
+# Usage:
+# run_suite_exe.sh [--g3=on|off] [--g4=on|off] [--garfield=on|off] [--examples="E01 E03 ..."]
 #
 # By I. Hrivnacova, IPN Orsay
 
 CURDIR=`pwd`
-OUTDIR=$CURDIR/log_run_exe
-G3EXEDIR=$CURDIR/../../geant3_install/bin
-G4EXEDIR=$CURDIR/../../geant4_vmc_install/bin
+OUTDIR=$CURDIR/logs/run_exe
 
 # Set 1 to 0 if you want to skip given MC or Garfield test
 TESTG3="1"
 TESTG4="1"
 
-# Run Garfield only with Root 5
+# Run Garfield optionally
 TESTGARFIELD="1"
-ROOT_VERSION=`root-config --version`
-if [ ${ROOT_VERSION:0:1} = "6" ]; then
-  TESTGARFIELD="0"
-fi
+
+# The default list of examples (all)
+ALL_EXAMPLES="E01 E02 E03 E06 A01 ExGarfield Gflash Monopole TR"
+EXAMPLES="$ALL_EXAMPLES"
+
+# Function arguments:
+# {1}  Run macro:  g3[q,b], g4[a,b]
+function run_mc_exe()
+{
+  echo "... Running ${1}vmc_example$EXAMPLE"
+  ${1}vmc_example$EXAMPLE >& $OUT/${1}vmc_example$EXAMPLE.out
+}
+
+
+# Process script arguments
+for arg in "${@}"
+do
+  #echo "got: $arg"
+  case $arg in
+    "--g3=on"        ) TESTG3="1" ;;
+    "--g3=off"       ) TESTG3="0" ;;
+    "--g4=on"        ) TESTG4="1" ;;
+    "--g4=off"       ) TESTG4="0" ;;
+    "--garfield=on"  ) TESTGARFIELD="1" ;;
+    "--garfield=off" ) TESTGARFIELD="0" ;;
+     --examples=*    ) EXAMPLES=${arg#--examples=} ;;
+    *                ) echo "Unsupported option $arg chosen."
+                       echo "run_suite_exe.sh [--g3=on|off] [--g4=on|off] [--garfield=on|off] [--examples=\"E01 E03 ...\"]"
+                       exit 1
+                       ;;
+  esac
+done
 
 # Recreate log directory only if running test for both G3 and G4
-if [ "$TESTG3" = "1" -a  "$TESTG4" = "1" ]; then
+if [ "$TESTG3" = "1" -a  "$TESTG4" = "1"  -a "$EXAMPLES" = "$ALL_EXAMPLES" ]; then
   rm -fr $OUTDIR
 fi  
 
-for EXAMPLE in E01 E02 E03 E06 A01 ExGarfield Gflash TR
+for EXAMPLE in $EXAMPLES
 do
   OUT=$OUTDIR/$EXAMPLE
   if [ ! -d $OUT ]; then
@@ -52,16 +79,22 @@ do
 
   cd $CURDIR/$EXAMPLE
   
-  if [ "$TESTG3" = "1" ]; then
-    EXE=$G3EXEDIR"/g3vmc_example"$EXAMPLE
-    echo "... Running g3vmc_example$EXAMPLE"
-    $EXE >& $OUT/g3vmc_example$EXAMPLE.out
+  if [ "$TESTG3" = "1" -a "$EXAMPLE" != "Monopole" ]; then
+    if [ "$EXAMPLE" = "E03" ]; then
+      run_mc_exe g3a
+      run_mc_exe g3b
+    else
+      run_mc_exe g3
+    fi
   fi  
 
   if [ "$TESTG4" = "1" ]; then
-    EXE=$G4EXEDIR"/g4vmc_example"$EXAMPLE
-    echo "... Running g4vmc_example$EXAMPLE"
-    $EXE >& $OUT/g4vmc_example$EXAMPLE.out
+    if [ "$EXAMPLE" = "E03" ]; then
+      run_mc_exe g4a
+      run_mc_exe g4b
+    else
+      run_mc_exe g4
+    fi
   fi  
   echo " "
