+++ examples/test_suite_exe.sh
@@ -1,4 +1,4 @@
-#!/bin/sh 
+#!/bin/bash
 #------------------------------------------------
 # The Virtual Monte Carlo examples
 # Copyright (C) 2014 Ivana Hrivnacova
@@ -10,10 +10,8 @@
 
 #
 # Run tests for all VMC examples from built executables and regenerate output
-# files. Only g4Config.C and g3tgeoConfig.C configurations are tested.
-# The executables are supposed to be found in
-# examples_build_g3/exampleName and examples_build_g4/exampleName
-# directories.
+# files. The executables are supposed to be found on the executable path or in the
+# path provided via --builddir option.
 #
 # Usage:
 # test_suite_exe.sh [--g3=on|off] [--g4=on|off] [--builddir=dir]
@@ -21,9 +19,8 @@
 # By I. Hrivnacova, IPN Orsay
 
 CURDIR=`pwd`
-OUTDIR=$CURDIR/log_exe
-G3EXEDIR=$CURDIR/../../g3examples_install/bin
-G4EXEDIR=$CURDIR/../../geant4_vmc_install/bin
+OUTDIR=$CURDIR/logs/test_exe
+EXEDIR=""
 PASSED="0"
 FAILED="0"
 
@@ -34,10 +31,40 @@ BUILDDIR=""
 
 # Run Garfield only with Root 5
 TESTGARFIELD="1"
-ROOT_VERSION=`root-config --version`
-if [ ${ROOT_VERSION:0:1} = "6" ]; then
-  TESTGARFIELD="0"
-fi
+
+# The default list of examples (all)
+ALL_EXAMPLES="E01 E02 E03 E06 A01 ExGarfield Gflash Monopole TR"
+EXAMPLES="$ALL_EXAMPLES"
+
+function print_help()
+{
+  echo "Usage:"
+  echo "test_suite_exe.sh [--g3=on|off] [--g4=on|off] [--garfield=on|off] [--examples=\"E01 E03 ...\"] [--builddir=dir] [--help|-h]"
+}
+
+STREAM_START_RED="\033[0;31m"
+STREAM_START_GREEN="\033[0;32m"
+STREAM_END_COLOR="\033[0m"
+
+# Function arguments:
+# {1} : message
+function start_test()
+{
+  echo -n ${1}
+}
+
+# Function arguments:
+# {1} : test result
+function evaluate_test()
+{
+  if [ "${1}" -ne "0" ]; then
+    echo -e "... ${STREAM_START_RED}failed${STREAM_END_COLOR}"
+    FAILED=`expr $FAILED + 1`
+  else
+    PASSED=`expr $PASSED + 1`
+    echo -e "... ${STREAM_START_GREEN}passed${STREAM_END_COLOR}"
+  fi
+}
 
 # Process script arguments
 for arg in "${@}"
@@ -50,17 +77,20 @@ do
     "--g4=off" ) TESTG4="0" ;;
     "--garfield=on"  ) TESTGARFIELD="1" ;;
     "--garfield=off" ) TESTGARFIELD="0" ;;
-     --builddir=* ) BUILDDIR=${arg#--builddir=} ;;
+     --examples=*    ) EXAMPLES=${arg#--examples=} ;;
+     --builddir=*    ) BUILDDIR=${arg#--builddir=} ;;
+    "--help" | "-h"  ) print_help
+                       exit 0
+                       ;;
     * ) echo "Unsupported option $arg chosen."
-        echo "Usage:"
-        echo "test_suite_exe.sh [--g3=on|off] [--g4=on|off] [--builddir=dir]"
+        print_help
         exit 1
         ;;
   esac
 done
 
 # Recreate log directory only if running test for both G3 and G4
-if [ "$TESTG3" = "1" -a  "$TESTG4" = "1" ]; then
+if [ "$TESTG3" = "1" -a  "$TESTG4" = "1" -a "$EXAMPLES" = "$ALL_EXAMPLES" ]; then
   rm -fr $OUTDIR
 fi
 
@@ -70,7 +100,7 @@ if [ "x${BUILDDIR}" != "x" ]; then
   export LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:${LIBS_FROM_BUILDDIR}
 fi
 
-for EXAMPLE in E01 E02 E03 E06 A01 ExGarfield Gflash TR
+for EXAMPLE in $EXAMPLES
 do
   OUT=$OUTDIR/$EXAMPLE
   if [ ! -d $OUT ]; then
@@ -85,71 +115,69 @@ do
   cd $CURDIR/$EXAMPLE
 
   if [ "x${BUILDDIR}" != "x" ]; then
-    G3EXEDIR=${BUILDDIR}/examples/$EXAMPLE
-    G4EXEDIR=${BUILDDIR}/examples/$EXAMPLE
+    EXEDIR=${BUILDDIR}/examples/$EXAMPLE/
   fi
 
   echo "... Example $EXAMPLE"
   
   if [ "$EXAMPLE" = "E01" -o "$EXAMPLE" = "E02" -o "$EXAMPLE" = "E06" ]; then
     if [ "$TESTG3" = "1" ]; then
-      EXE=$G3EXEDIR"/g3vmc_test"$EXAMPLE
-      echo "... Running test with G3, geometry via TGeo, TGeo navigation"
+      EXE=$EXEDIR"g3vmc_test"$EXAMPLE
+      start_test "... Running test with G3, geometry via TGeo, TGeo navigation"
       TMP_FAILED="0"
       $EXE -g3g TGeant3TGeo -rm "test_$EXAMPLE.C(\"\", kFALSE)" >& $OUT/test_g3_tgeo_tgeo.out
       if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
       # stack popper test
       if [ "$EXAMPLE" = "E06" ]; then
-        #echo "... Running test with G3, geometry via TGeo, TGeo navigation, with Feedback photons" 
         $EXE -g3g TGeant3TGeo -rm "test_E06_2.C(\"\", kFALSE)"  >& tmpfile
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
         cat tmpfile >> $OUT/test_g3_tgeo_tgeo.out
       fi
-      if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$TMP_FAILED"
 
-      echo "... Running test with G3, geometry via VMC,  Native navigation"
+      start_test "... Running test with G3, geometry via VMC,  Native navigation"
       $EXE -g3g TGeant3 -rm "test_$EXAMPLE.C(\"\", kTRUE)" >& $OUT/test_g3_vmc_nat.out
-      if [ "$?" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$?"
 
-      echo "... Running test with G3, geometry via VMC,  TGeo navigation"
+      start_test "... Running test with G3, geometry via VMC,  TGeo navigation"
       $EXE -g3g TGeant3TGeo -rm "test_$EXAMPLE.C(\"\", kTRUE)" >& $OUT/test_g3_vmc_tgeo.out
-      if [ "$?" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$?"
     fi  
     if [ "$TESTG4" = "1" ]; then
-      EXE=$G4EXEDIR"/g4vmc_test"$EXAMPLE
+      EXE=$EXEDIR"g4vmc_test"$EXAMPLE
       TMP_FAILED="0"
-      echo "... Running test with G4, geometry via TGeo, Native navigation"
+      start_test "... Running test with G4, geometry via TGeo, Native navigation"
       $EXE -g4g geomRootToGeant4 -g4vm "" -rm "test_$EXAMPLE.C(\"\", kFALSE)" >& $OUT/test_g4_tgeo_nat.out
       if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
       # stack popper test
       if [ "$EXAMPLE" = "E06" ]; then
-        #echo "... Running test with G4, geometry via TGeo, Native navigation, with stackPopper" 
+        #start_test "... Running test with G4, geometry via TGeo, Native navigation, with stackPopper"
         $EXE -g4g geomRootToGeant4 -g4vm "" -g4pl "FTFP_BERT_EMV+optical" -g4sp stackPopper -rm "test_E06_2.C(\"\", kFALSE)" >& tmpfile
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
         cat tmpfile >> $OUT/test_g4_tgeo_nat.out
       fi
-      if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$TMP_FAILED"
 
       TMP_FAILED="0"
-      echo "... Running test with G4, geometry via TGeo, TGeo navigation"
+      start_test "... Running test with G4, geometry via TGeo, TGeo navigation"
       $EXE -g4g geomRoot -g4vm "" -rm "test_$EXAMPLE.C(\"\", kFALSE)" >& $OUT/test_g4_tgeo_tgeo.out
       if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
       # stack popper test
       if [ "$EXAMPLE" = "E06" ]; then
-        #echo "... Running test with G4, geometry via TGeo, Native navigation, with stackPopper" 
+        #start_test "... Running test with G4, geometry via TGeo, Native navigation, with stackPopper"
         $EXE -g4g geomRoot -g4vm "" -g4pl "FTFP_BERT_EMV+optical" -g4sp stackPopper -rm "test_E06_2.C(\"\", kFALSE)" >& tmpfile
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
         cat tmpfile >> $OUT/test_g4_tgeo_tgeo.out
       fi
-      if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$TMP_FAILED"
 
-      echo "... Running test with G4, geometry via VMC,  Native navigation"
+      start_test "... Running test with G4, geometry via VMC,  Native navigation"
       $EXE -g4g geomVMCtoGeant4 -g4vm "" -rm "test_$EXAMPLE.C(\"\", kTRUE)" >& $OUT/test_g4_vmc_nat.out
-      if [ "$?" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$?"
 
-      echo "... Running test with G4, geometry via VMC,  TGeo navigation"
+      start_test "... Running test with G4, geometry via VMC,  TGeo navigation"
       $EXE -g4g geomVMCtoRoot -g4vm "" -rm "test_$EXAMPLE.C(\"\", kTRUE)" >& $OUT/test_g4_vmc_tgeo.out
-      if [ "$?" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$?"
     fi  
   fi  
 
@@ -162,12 +190,12 @@ do
       fi
 
       cd $CURDIR/$EXAMPLE
-      echo "... Example $EXAMPLE/$OPTION"
+      start_test "... Example $EXAMPLE/$OPTION"
 
       # Run all macros + special configuration available only in E03 test
       if [ "$TESTG3" = "1" ]; then
-        EXE=$G3EXEDIR"/g3vmc_test"$OPTION
-        echo "... Running test with G3, geometry via TGeo, TGeo navigation"
+        EXE=$EXEDIR"g3vmc_test"$OPTION
+        start_test "... Running test with G3, geometry via TGeo, TGeo navigation"
         TMP_FAILED="0"
         $EXE -g3g TGeant3TGeo -rm "test_E03_1.C(\"\", kFALSE)" >& $OUT/test_g3_tgeo_tgeo.out
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
@@ -183,9 +211,9 @@ do
         $EXE -g3g TGeant3TGeo -rm "test_E03_5.C(\"\", kFALSE)" >& tmpfile
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
         cat tmpfile >> $OUT/test_g3_tgeo_tgeo.out
-        if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+        evaluate_test "$TMP_FAILED"
 
-        echo "... Running test with G3, geometry via VMC,  Native navigation"
+        start_test "... Running test with G3, geometry via VMC,  Native navigation"
         TMP_FAILED="0"
         $EXE -g3g TGeant3 -rm "test_E03_1.C(\"\", kTRUE)" >& $OUT/test_g3_vmc_nat.out
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
@@ -201,9 +229,9 @@ do
         $EXE -g3g TGeant3 -rm "test_E03_5.C(\"\", kTRUE)" >& tmpfile
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
         cat tmpfile >> $OUT/test_g3_vmc_nat.out
-        if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+        evaluate_test "$TMP_FAILED"
 
-        echo "... Running test with G3, geometry via VMC,  TGeo navigation"
+        start_test "... Running test with G3, geometry via VMC,  TGeo navigation"
         TMP_FAILED="0"
         $EXE -g3g TGeant3TGeo -rm "test_E03_1.C(\"\", kTRUE)" >& $OUT/test_g3_vmc_tgeo.out
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
@@ -219,12 +247,12 @@ do
         $EXE -g3g TGeant3TGeo -rm "test_E03_5.C(\"\", kTRUE)" >& tmpfile
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
         cat tmpfile >> $OUT/test_g3_vmc_tgeo.out
-        if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+        evaluate_test "$TMP_FAILED"
       fi
       # Run all macros + special configuration available only in E03 test
       if [ "$TESTG4" = "1" ]; then
-        EXE=$G4EXEDIR"/g4vmc_test"$OPTION
-        echo "... Running test with G4, geometry via TGeo, Native navigation"
+        EXE=$EXEDIR"g4vmc_test"$OPTION
+        start_test "... Running test with G4, geometry via TGeo, Native navigation"
         TMP_FAILED="0"
         $EXE -g4g geomRootToGeant4 -g4vm "" -rm "test_E03_1.C(\"\", kFALSE)" >& $OUT/test_g4_tgeo_nat.out
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
@@ -243,9 +271,9 @@ do
         $EXE -g4g geomRootToGeant4 -g4uc "field" -g4vm "" -rm "test_E03_6.C(\"\", kFALSE)" >& tmpfile
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
         cat tmpfile >> $OUT/test_g4_tgeo_nat.out
-        if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+        evaluate_test "$TMP_FAILED"
 
-        echo "... Running test with G4, geometry via TGeo, TGeo navigation"
+        start_test "... Running test with G4, geometry via TGeo, TGeo navigation"
         TMP_FAILED="0"
         $EXE -g4g geomRoot -g4vm "" -rm "test_E03_1.C(\"\", kFALSE)" >& $OUT/test_g4_tgeo_tgeo.out
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
@@ -264,9 +292,9 @@ do
         $EXE -g4g geomRoot -g4uc "field" -g4vm "" -rm "test_E03_6.C(\"\", kFALSE)" >& tmpfile
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
         cat tmpfile >> $OUT/test_g4_tgeo_tgeo.out
-        if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+        evaluate_test "$TMP_FAILED"
 
-        echo "... Running test with G4, geometry via VMC,  Native navigation"
+        start_test "... Running test with G4, geometry via VMC,  Native navigation"
         TMP_FAILED="0"
         $EXE -g4g geomVMCtoGeant4 -g4vm "" -rm "test_E03_1.C(\"\", kTRUE)" >& $OUT/test_g4_vmc_nat.out
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
@@ -279,9 +307,9 @@ do
         $EXE -g4g geomVMCtoGeant4 -g4vm "" -rm "test_E03_4.C(\"\", kTRUE)" >& tmpfile
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
         cat tmpfile >> $OUT/test_g4_vmc_nat.out
-        if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+        evaluate_test "$TMP_FAILED"
 
-        echo "... Running test with G4, geometry via VMC,  TGeo navigation"
+        start_test "... Running test with G4, geometry via VMC,  TGeo navigation"
         TMP_FAILED="0"
         $EXE -g4g geomVMCtoRoot -g4vm "" -rm "test_E03_1.C(\"\", kTRUE)" >& $OUT/test_g4_vmc_tgeo.out
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
@@ -294,9 +322,9 @@ do
         $EXE -g4g geomVMCtoRoot -g4vm "" -rm "test_E03_4.C(\"\", kTRUE)" >& tmpfile
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
         cat tmpfile >> $OUT/test_g4_vmc_tgeo.out
-        if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+        evaluate_test "$TMP_FAILED"
 
-        echo "... Running test with G4, geometry via G4,   Native navigation"
+        start_test "... Running test with G4, geometry via G4,   Native navigation"
         TMP_FAILED="0"
         $EXE -g4uc geometry -g4vm "" -rm "test_E03_1.C(\"\", kFALSE)" >& $OUT/test_g4_g4_nat.out
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
@@ -309,9 +337,9 @@ do
         $EXE -g4uc geometry -g4vm "" -rm "test_E03_4.C(\"\", kFALSE)" >& tmpfile
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
         cat tmpfile >> $OUT/test_g4_g4_nat.out
-        if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+        evaluate_test "$TMP_FAILED"
 
-        echo "... Running test with G4,  geometry via TGeo, Native navigation, User physics list"
+        start_test "... Running test with G4,  geometry via TGeo, Native navigation, User physics list"
         TMP_FAILED="0"
         $EXE -g4uc physics-list -g4vm "" -rm "test_E03_1.C(\"\", kFALSE)" >& $OUT/test_g4_g4_nat_pl.out
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
@@ -324,21 +352,21 @@ do
         $EXE -g4uc physics-list -g4vm "" -rm "test_E03_4.C(\"\", kFALSE)" >& tmpfile
         if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
         cat tmpfile >> $OUT/test_g4_g4_nat_pl.out
-        if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+        evaluate_test "$TMP_FAILED"
       fi
     done
   fi
 
   if [ "$EXAMPLE" = "A01" ]; then 
     if [ "$TESTG3" = "1" ]; then
-      EXE=$G3EXEDIR"/g3vmc_test"$EXAMPLE
-      echo "... Running test with G3, geometry via TGeo, TGeo navigation"
+      EXE=$EXEDIR"g3vmc_test"$EXAMPLE
+      start_test "... Running test with G3, geometry via TGeo, TGeo navigation"
       $EXE -g3g TGeant3TGeo -rm "test_A01_1.C(\"\", kFALSE)" >& $OUT/test_g3_tgeo_tgeo.out
-      if [ "$?" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$?"
     fi  
     if [ "$TESTG4" = "1" ]; then
-      EXE=$G4EXEDIR"/g4vmc_test"$EXAMPLE
-      echo "... Running test with G4, geometry via TGeo, Native navigation"
+      EXE=$EXEDIR"g4vmc_test"$EXAMPLE
+      start_test "... Running test with G4, geometry via TGeo, Native navigation"
       TMP_FAILED="0"
       $EXE -g4g geomRootToGeant4 -g4vm "" -rm "test_A01_1.C(\"\", kFALSE)" >& $OUT/test_g4_tgeo_nat.out
       if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
@@ -346,9 +374,9 @@ do
       if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
       cat tmpfile >> $OUT/test_g4_tgeo_nat.out
       rm tmpfile
-      if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$TMP_FAILED"
 
-      echo "... Running test with G4, geometry via TGeo, TGeo navigation"
+      start_test "... Running test with G4, geometry via TGeo, TGeo navigation"
       TMP_FAILED="0"
       $EXE -g4g geomRoot -g4vm "" -rm "test_A01_1.C(\"\", kFALSE)" >& $OUT/test_g4_tgeo_tgeo.out
       if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
@@ -356,9 +384,9 @@ do
       if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
       cat tmpfile >> $OUT/test_g4_tgeo_tgeo.out
       rm tmpfile
-      if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$TMP_FAILED"
 
-      echo "... Running test with G4, geometry via G4, Native navigation"
+      start_test "... Running test with G4, geometry via G4, Native navigation"
       TMP_FAILED="0"
       $EXE -g4uc geometry -g4vm "" -rm "test_A01_1.C(\"\", kFALSE)" >& $OUT/test_g4_g4_nat.out
       if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
@@ -366,38 +394,34 @@ do
       if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
       cat tmpfile >> $OUT/test_g4_g4_nat.out
       rm tmpfile
-      if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$TMP_FAILED"
     fi  
   fi  
 
   # new examples (without old geomtry definition)
   #
-  if [ "$EXAMPLE" = "ExGarfield" -o "$EXAMPLE" = "Gflash" -o "$EXAMPLE" = "TR" ]; then
+  if [ "$EXAMPLE" = "ExGarfield" -o "$EXAMPLE" = "Gflash" -o "$EXAMPLE" = "Monopole" -o "$EXAMPLE" = "TR" ]; then
 
-    if [ "$TESTG3" = "1" ]; then
-      EXE=$G3EXEDIR"/g3vmc_test"$EXAMPLE
-      echo "... Running test with G3, geometry via TGeo, TGeo navigation"
-      TMP_FAILED="0"
+    if [ "$TESTG3" = "1" -a "$EXAMPLE" != "ExGarfield" -a  "$EXAMPLE" != "Monopole" ]; then
+      EXE=$EXEDIR"g3vmc_test"$EXAMPLE
+      start_test "... Running test with G3, geometry via TGeo, TGeo navigation"
       $EXE -g3g TGeant3TGeo -rm "test_$EXAMPLE.C(\"\")" >& $OUT/test_g3_tgeo_tgeo.out
-      if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
-      if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$?"
     fi
 
     if [ "$TESTG4" = "1" ]; then
-      EXE=$G4EXEDIR"/g4vmc_test"$EXAMPLE
-      echo "... Running test with G4, geometry via TGeo, Native navigation"
-      TMP_FAILED="0"
+      EXE=$EXEDIR"g4vmc_test"$EXAMPLE
+      start_test "... Running test with G4, geometry via TGeo, Native navigation"
       $EXE -g4g geomRootToGeant4 -g4vm "" -rm "test_$EXAMPLE.C(\"\")" >& $OUT/test_g4_tgeo_nat.out
       if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
-      if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$?"
 
-      echo "... Running test with G4, geometry via TGeo, TGeo navigation"
+      start_test "... Running test with G4, geometry via TGeo, TGeo navigation"
       TMP_FAILED="0"
       $EXE -g4g geomRoot -g4vm "" -rm "test_$EXAMPLE.C(\"\")" >& $OUT/test_g4_tgeo_tgeo.out
       if [ "$?" -ne "0" ]; then TMP_FAILED="1" ; fi
-      if [ "$TMP_FAILED" -ne "0" ]; then FAILED=`expr $FAILED + 1`; else PASSED=`expr $PASSED + 1`; fi
+      evaluate_test "$?"
     fi
-
   fi
 
   echo " "
